fixed #10978 solved
